Move TopAddButtonAndDialog out of listTable folder
